[contactsd] Also tokenise SIM contact name
Previously, only display labels received from XMPP contacts were
tokenised (via locale-specific heuristics) into structured name
fields.  This commit hoists that logic into a util file and uses
that from both the SIM and XMPP (telepathy) plugins.
