+++ tests/ut_simplugin/ut_simplugin.pro
@@ -18,10 +18,12 @@ INCLUDEPATH += \
 
 HEADERS += \
     test-sim-plugin.h \
-    ../../plugins/sim/cdsimcontroller.h
+    ../../plugins/sim/cdsimcontroller.h \
+    ../../src/util.h
 
 SOURCES += \
     test-sim-plugin.cpp \
-    ../../plugins/sim/cdsimcontroller.cpp
+    ../../plugins/sim/cdsimcontroller.cpp \
+    ../../src/util.cpp
 
 INSTALLS += target
